+++ docs/api/version-history.md
@@ -74,6 +74,8 @@ keywords: "API, Docker, rcli, REST, documentation"
   job-mode service.
 * `GET /containers/{id}/stats` now accepts a query param (`one-shot`) which, when used with `stream=false` fetches a
   single set of stats instead of waiting for two collection cycles to have 2 CPU stats over a 1 second period.
+* The `KernelMemory` field in `HostConfig.Resources` is now deprecated.
+* The `KernelMemory` field in `Info` is now deprecated.
 
 ## v1.40 API changes